refactor: extract dashboard routes into routes_dashboard.py blueprint
Move 11 routes plus shared helpers out of app.py into a new
routes_dashboard.py blueprint named "dashboard":

- Dashboard views: /, /tiled_dash, /compact_dash, /dbdump
- Settings page: /settings (backup, restore, version, users, groups)
- Group CRUD: /update_group, /add_group, /delete_group
- Service CRUD: /add, /edit/<id>
- Image proxy: /images/<filename>

Helpers also moved: the `flash_is_present` deduplicator used by the
restore handler.

Module-level state moved to app.config so blueprint routes can read
via current_app.config:
- IMAGE_DIR — /config/images, used by 3 routes plus background jobs
- CONFIG_FROM_ENV / CONFIG_FROM_FILE — passed to the /settings page
- VERSION_INFO — version.txt cached once at startup. The /settings
  page previously re-read /app/version.txt per request; now it
  reads the cached dict. version.txt is set at Docker build time
  and never changes at runtime, so caching is correct.

Blueprint endpoint names are now `dashboard.dashboard`,
`dashboard.tiled_dashboard`, etc. Updated url_for() call sites:
- routes_auth.py — 8 references (dashboard, settings)
- templates/dashboard.html — serve_image, edit_entry
- templates/edit_entry.html — edit_entry
- templates/compact_dash.html — serve_image
- templates/tiled_dash.html — serve_image
- templates/add_entry.html — add_entry
- templates/settings.html — settings, update_group, delete_group,
  add_group

Issue D2 in the PRD (grouping/sorting duplicated across the three
dashboard views) is deliberately NOT deduplicated here — Phase 4
plan calls this out as a follow-up.

Dropped from app.py imports: defaultdict, urlparse/urljoin,
joinedload, asc, desc, nullslast, send_file, send_from_directory,
make_response, flash, redirect, url_for, fetch_icon_if_missing
(still referenced by verify_and_fetch_missing_icons which moves in
Step 7), is_admin_required.
